chore: move ember specific impls to `migration-graph-ember`
- Rename `MigrationGraph` to `ProjectGraph` to align with taxonomy
  - Project has many Packages.
  - Packages have many modules (files).
- Moved `isPackageCovertedToTypescript` to `Package` in `migration-graph-shared`.
  - Added some tests.
- Moved `discover-ember-packages` to `migration-graph-ember`.
  - Updated usage in `cli`.
- Updated imports for usage of `ProjectGraph`
