+++ packages/migration-graph-ember/src/index.ts
@@ -19,4 +19,5 @@ export {
   getEmberAddonName,
 } from './utils/ember';
 
+export { discoverEmberPackages } from './utils/discover-ember-packages';
 export { discoverServiceDependencies } from './utils/discover-ember-service-dependencies';
