+++ packages/migration-graph-shared/src/entities/package.ts
@@ -2,6 +2,7 @@
 import path, { resolve } from 'path';
 import { readJsonSync, writeJsonSync } from 'fs-extra';
 import sortPackageJson from 'sort-package-json';
+import { sync as fastGlobSync } from 'fast-glob';
 
 import { removeNestedPropertyValue, setNestedPropertyValue } from '../utils/pojo';
 import { getWorkspaceGlobs } from '../utils/workspace';
@@ -227,5 +228,33 @@ export class Package {
     const sorted: Record<any, any> = sortPackageJson(this.packageJson);
     const pathToPackageJson = path.join(this.path, 'package.json');
     writeJsonSync(pathToPackageJson, sorted, { spaces: 2 });
+  }
+
+  isConvertedToTypescript(conversionLevel?: string): boolean {
+    const fastGlobConfig = {
+      absolute: true,
+      cwd: this.path,
+      ignore: ['**/node_modules/**'],
+    };
+    // ignore a tests directory if we only want to consider the source
+    if (conversionLevel === 'source-only') {
+      fastGlobConfig.ignore.push('**/tests/**');
+    }
+
+    // ignore some common .js files unless considering "full" conversion
+    if (conversionLevel !== 'full') {
+      this.excludePatterns;
+      fastGlobConfig.ignore.push(...this.excludePatterns);
+    }
+
+    // if there's a tsconfig
+    const hasTSConfig = fastGlobSync('tsconfig.json', fastGlobConfig);
+    // if there aren't any .js files in addon (minus the ignore list)
+    const hasJS = fastGlobSync('**/*.js', fastGlobConfig);
+
+    if (!!hasTSConfig?.length && !hasJS?.length) {
+      return true;
+    }
+    return false;
   }
 }
